+++ lib/Service/OpenProjectAPIService.php
@@ -246,10 +246,25 @@ class OpenProjectAPIService {
 	}
 
 	/**
-	 * wrapper around IURLGenerator::getBaseUrl() to make it easier to mock in tests
+	 * @return string
 	 */
-	public function getBaseUrl(): string {
-		return $this->config->getSystemValueString('overwrite.cli.url');
+	public function getNCBaseUrl(): string {
+		$message = 'Invalid or missing "overwrite.cli.url" system configuration.';
+		$ncUrl = $this->config->getSystemValueString('overwrite.cli.url');
+		if (!$ncUrl) {
+			$this->logger->error($message, ['app' => $this->appName]);
+			return '';
+		}
+
+		$parsedUrl = parse_url($ncUrl);
+		if (!$parsedUrl || !isset($parsedUrl['scheme']) || !isset($parsedUrl['host'])) {
+			$this->logger->error($message, ['app' => $this->appName]);
+			return '';
+		}
+		if ($parsedUrl['scheme'] !== 'https') {
+			$ncUrl = str_replace('http://', 'https://', $ncUrl);
+		}
+		return $ncUrl;
 	}
 
 	/**
@@ -303,7 +318,7 @@ class OpenProjectAPIService {
 		if ($onlyLinkableWorkPackages) {
 			$filters[] = [
 				'linkable_to_storage_url' =>
-					['operator' => '=', 'values' => [urlencode($this->getBaseUrl())]]
+					['operator' => '=', 'values' => [urlencode($this->getNCBaseUrl())]]
 			];
 		}
 
@@ -438,6 +453,15 @@ class OpenProjectAPIService {
 				$options['headers']['Content-Type'] = 'application/json';
 			}
 		}
+
+		$sanitizedOptions = $this->sanitizeReqOptionsForLog($options);
+		$requestLog = json_encode([
+			'method' => $method,
+			'url' => $url,
+			'headers' => $sanitizedOptions['headers'],
+			'body' => $sanitizedOptions['body'],
+		]);
+		$this->logger->debug('OpenProject API request: ' . $requestLog, ['app' => $this->appName]);
 
 		if ($method === 'GET') {
 			$response = $this->client->get($url, $options);
@@ -757,7 +781,7 @@ class OpenProjectAPIService {
 				],
 				'_links' => [
 					'storageUrl' => [
-						'href' => $this->getBaseUrl()
+						'href' => $this->getNCBaseUrl()
 					]
 				]
 			];
@@ -1511,7 +1535,7 @@ class OpenProjectAPIService {
 		}
 		$filters[] = [
 			'storageUrl' =>
-				['operator' => '=', 'values' => [$this->getBaseUrl()]],
+				['operator' => '=', 'values' => [$this->getNCBaseUrl()]],
 			'userAction' =>
 				['operator' => '&=', 'values' => ["file_links/manage", "work_packages/create"]]
 		];
@@ -1866,5 +1890,43 @@ class OpenProjectAPIService {
 		}
 
 		return $appInfo['name'];
+	}
+
+	/**
+	 * @param array $options
+	 *
+	 * @return array
+	 */
+	public function sanitizeReqOptionsForLog(array $options): array {
+		$sanitizedOptions = [
+			'headers' => [],
+			'body' => null,
+		];
+		$headers = ['authorization', 'cookie', 'set-cookie'];
+		if (isset($options['headers'])) {
+			foreach ($options['headers'] as $key => $value) {
+				if (in_array(strtolower($key), $headers)) {
+					$sanitizedOptions['headers'][$key] = '[REDACTED]';
+				} else {
+					$sanitizedOptions['headers'][$key] = $value;
+				}
+			}
+		}
+
+		if (isset($options['body'])) {
+			try {
+				$body = json_decode($options['body'], true, 512, JSON_THROW_ON_ERROR);
+				$fields = ['password', 'client_secret', 'refresh_token', 'access_token', 'token'];
+				foreach ($fields as $field) {
+					if (isset($body[$field])) {
+						$body[$field] = '[REDACTED]';
+					}
+				}
+				$sanitizedOptions['body'] = $body;
+			} catch (\JsonException) {
+				$sanitizedOptions['body'] = '[DATA]';
+			}
+		}
+		return $sanitizedOptions;
 	}
 }